backend/requirements/base.txt
Django>=5.2,<5.3
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.5,<5.6
django-cors-headers>=4.7,<4.8
psycopg[binary]>=3.2,<3.3
celery>=5.5,<5.6
redis>=6.0,<6.1
pgvector>=0.4,<0.5
django-celery-results>=2.6,<2.7
gunicorn>=23.0,<24.0
